replace .texi file with plain text
New texi2dvi 6.1 breaks R and vegan FAQ-vegan.texi failed. The
quick fix is to replace the .texi file with plain text and remove
vignettes/Makefile. The main (devel) branch will translate the .texi
to .Rmd, but that fix would conflict with the 2.3 tree and we go here
only for this quick fix.